chore: revert the PaginationBar rename and move to ensure 100% compatibility
