+++ packages/eui/src/components/basic_table/index.ts
@@ -24,5 +24,6 @@ export type {
   EuiTableSelectionType,
   EuiTableSortingType,
 } from './table_types';
+export type { Pagination } from './pagination_bar';
 export type { DefaultItemAction, CustomItemAction } from './action_types';
 export { useEuiBasicTablePanelProps } from './use_panel_props';
