--- packages/eui/src/components/basic_table/pagination_bar/index.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-
-export {
-  EuiBasicTablePaginationBar,
-  type EuiBasicTablePaginationBarPagination,
-} from './pagination_bar';
